revert KimiGDNStateMixin refactor
